requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.6
packaging==24.1
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
sqlparse==0.5.0
uritemplate==4.1.1
